refactor: 파일 이름에 오타를 수정했다.
