+++ horror-alram-front/src/components/AlarmSwitch.tsx
